[AC-9270] move functions from expert profile to core profile
