+++ accelerator/models/core_profile.py
@@ -1,6 +1,9 @@
-from datetime import timedelta
-
 import swapper
+from pytz import utc
+from datetime import (
+    datetime,
+    timedelta,
+)
 from django.db.models import (
     OuterRef,
     Q,
@@ -21,6 +24,9 @@ from accelerator_abstract.models import (
     UPCOMING_PROGRAM_STATUS,
 )
 from accelerator_abstract.models.base_core_profile import BaseCoreProfile
+from accelerator_abstract.models.base_judging_round import (
+    ONLINE_JUDGING_ROUND_TYPE,
+)
 from accelerator_abstract.models.base_user_utils import is_employee
 
 SHORT_BIO_MAX_LENGTH = 140
@@ -98,6 +104,10 @@ class CoreProfile(BaseCoreProfile, PolymorphicModel):
     def mentor_profile_url(self):
         if self.is_mentor():
             return reverse('mentor_view', args=(self.user.id,))
+
+    @property
+    def active_round(self):
+        return self.active_judging_rounds().first()
 
     def is_confirmed_mentor(self):
         return self.user.programrolegrant_set.filter(
@@ -234,6 +244,148 @@ class CoreProfile(BaseCoreProfile, PolymorphicModel):
             program_family=OuterRef('pk'),
             program_status__in=[ACTIVE_PROGRAM_STATUS, ENDED_PROGRAM_STATUS]
         ).order_by("-created_at").values('pk')[:1]
+
+    def get_stats(self,
+                  round_active=True,
+                  judging_round=None):
+        active_online_rounds = self.active_judging_rounds().filter(
+            round_type=ONLINE_JUDGING_ROUND_TYPE)
+
+        if not judging_round and (
+                round_active and
+                not active_online_rounds.exists()
+        ):
+            return {}
+
+        assignment_counts = self.get_assignment_counts(
+            round_active=round_active,
+            judging_round=judging_round)
+        stats = {
+            'C': self.get_capacity(judging_round),
+            'A': assignment_counts[0],
+            'Q': min(
+                assignment_counts[0], self.get_current_quota(judging_round)),
+            'deadline': self.get_deadline(judging_round),
+            'S': assignment_counts[1],
+            'P': self.get_all_panel_assignments()
+        }
+        stats['due_now'] = max((stats['Q'] - stats['S']), 0)
+        stats['additional_due'] = max(
+            (max(stats['C'], stats['A']) - stats['S']) - stats['due_now'], 0)
+        return stats
+
+    def active_judging_rounds(self):
+        JudgingRound = swapper.load_model('accelerator', 'JudgingRound')
+        judge_grants = self.user.programrolegrant_set.filter(
+            program_role__user_role__name=UserRole.JUDGE)
+        program_ids = judge_grants.values_list('program_role__program_id')
+        rounds = JudgingRound.objects.filter(
+            is_active=True,
+            program_id__in=program_ids)
+        return rounds
+
+    def get_all_panel_assignments(self,
+                                  round_active=True,
+                                  future_only=False,
+                                  judging_round=None):
+        jpa_filter = self._create_jpa_filter(round_active,
+                                             judging_round)
+        jpas = self.user.judgepanelassignment_set.filter(
+            **jpa_filter).order_by('panel__panel_time__start_date_time')
+        if future_only:
+            now = utc.localize(datetime.today())
+            jpas = jpas.filter(panel__panel_time__start_date_time__gt=now)
+        return jpas
+
+    def _create_jpa_filter(self, round_active, judging_round):
+        # It would be better to somehow move this to JudgePanelAssignment or
+        # a JudgePanelAssignmentManager object such that it can still
+        # be accessed through self.judgepanelassignment_set, but I'm
+        # not sure the best way to do that.
+        filter = {
+            'judge': self.user,
+            'scenario__is_active': True
+        }
+        if judging_round:
+            filter['scenario__judging_round'] = judging_round
+            return filter
+        if round_active:
+            filter['scenario__judging_round__is_active__exact'] = True
+        active_rounds = self.active_judging_rounds()
+        if active_rounds.exists():
+            filter['scenario__judging_round__in'] = active_rounds
+        return filter
+
+    def get_deadline(self, judging_round):
+        if judging_round and judging_round.end_date_time:
+            return judging_round.end_date_time
+        else:
+            return "N/A"
+
+    def get_current_quota(self, judging_round):
+        current_quota = 0
+        commitment = self.get_commitment(judging_round)
+        try:
+            current_quota = commitment.current_quota or 0
+        except AttributeError:
+            pass
+        finally:
+            return current_quota
+
+    def get_commitment(self, judging_round):
+        if judging_round:
+            return self.user.judgeroundcommitment_set.filter(
+                judging_round=judging_round).first()
+        else:
+            return None
+
+    def get_capacity(self, judging_round):
+        """return the committed capacity of this judge for this active round
+        """
+        capacity = 0
+        commitment = self.get_commitment(judging_round)
+        try:
+            capacity = commitment.capacity or 0
+        except AttributeError:
+            pass
+        finally:
+            return capacity
+
+    def get_assignment_counts(self, round_active=True,
+                              judging_round=None):
+        """return the count of all applications and completed applications
+        """
+        user = self.user
+        app_count = completed_count = 0
+        panel_assignments = self.get_all_panel_assignments(
+            round_active=round_active,
+            judging_round=judging_round)
+        for assignment in panel_assignments:
+            assignment_completed = assignment.assignment_status == 'COMPLETE'
+            panel = assignment.panel
+            applications = panel.applications.all()
+            forms = panel.get_feedback_forms_for_user(user)
+            for application in applications:
+                app_count += 1
+                if assignment_completed:
+                    form = forms.get(str(application.pk), None)
+                    if form and form.feedback_status != 'INCOMPLETE':
+                        completed_count += 1
+        return app_count, completed_count
+
+    def is_judge(self, program=None, state=None):
+        """If program is specified, is the user a judge in that program.
+        Otherwise, is the user a judge in any program."""
+        role_dict = {'desired': [UserRole.DESIRED_JUDGE],
+                     'confirmed': [UserRole.JUDGE]}
+        roles = role_dict.get(state, [UserRole.DESIRED_JUDGE, UserRole.JUDGE])
+        if program:
+            return self.user.programrolegrant_set.filter(
+                program_role__program__exact=program,
+                program_role__user_role__name__in=roles).exists()
+        else:
+            return self.user.programrolegrant_set.filter(
+                program_role__user_role__name__in=roles).exists()
 
 
 def _get_office_hour_holder_active_programs(user):
